Dask interface: Accept and forward the new `if-exists` query param
The new `if-exists` query parameter will be forwarded to Dask's
`to_sql()` method. By default, `influxio` will use `fail`.

When targeting the SQLAlchemy database interface, the target table will
be created automatically, if it does not exist. The `if-exists` URL
query parameter can be used to configure this behavior. The default
value is `fail`.

* fail: Raise a ValueError.
* replace: Drop the table before inserting new values.
* append: Insert new values to the existing table.